+++ src/finance_geometry.cc
@@ -18,19 +18,19 @@ void geometry::Geometry::SetOuterEdges(double price_quote,
 
 }
 
-int geometry::Geometry::GetOuterRedColor() {
+double geometry::Geometry::GetOuterRedColor() {
   return this->outer_red_value;
 }
 
-int geometry::Geometry::GetOuterGreenColor() {
+double geometry::Geometry::GetOuterGreenColor() {
   return this->outer_green_value;
 }
 
-int geometry::Geometry::GetOuterBlueColor() {
+double geometry::Geometry::GetOuterBlueColor() {
   return this->outer_blue_value;
 }
 
-int geometry::Geometry::GetOuterEdgeNumber() {
+double geometry::Geometry::GetOuterEdgeNumber() {
   return this->outer_edge_number;
 }
 
@@ -67,19 +67,19 @@ void geometry::Geometry::SetInnerEdges(int buy_rec,
 
 }
 
-int geometry::Geometry::GetInnerRedColor() {
+double geometry::Geometry::GetInnerRedColor() {
   return this->inner_red_value;
 }
 
-int geometry::Geometry::GetInnerGreenColor() {
+double geometry::Geometry::GetInnerGreenColor() {
   return this->inner_green_value;
 }
 
-int geometry::Geometry::GetInnerBlueColor() {
+double geometry::Geometry::GetInnerBlueColor() {
   return this->inner_blue_value;
 }
 
-int geometry::Geometry::GetInnerEdgeNumber() {
+double geometry::Geometry::GetInnerEdgeNumber() {
   return this->inner_edge_number;
 }
